# docs/api.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: "Fast & Furious cinema"
  description: "Backend API documentation for Fast & Furious cinema"
host: "api.example.com"
basePath: "/api"
schemes:
  - "http"
securityDefinitions:
  basicAuth:
    type: basic
paths:
...
  /showtimes:
    post:
      summary: "Create movie showtime"
      security:
        - basicAuth: []
      parameters:
        - in: body
          name: CreateShowtimeRequest
          schema:
            $ref: "#/definitions/CreateShowtimeRequest"
      responses:
        201:
          description: "Showtime created"
          schema:
            $ref: "#/definitions/ShowtimeResponse"
        401:
          description: "Client is not authorized"
    get:
      summary: "Get a list of showtimes"
      security:
        - basicAuth: []
      parameters:
        - in: query
          name: movie_id
          type: string
          format: uuid
        - in: query
          name: date_start_gte
          type: string
          format: date-time
        - in: query
          name: date_start_lte
          type: string
          format: date-time
      responses:
        200:
          description: "List of showtimes that match the given criteria or an empty list if no matching items are found"
          schema:
            type: array
            items:
              $ref: "#/definitions/ShowtimeResponse"
        401:
          description: "Client is not authorized"
...
definitions:
...
  Price:
    type: object
    properties:
      amount:
        type: string
        format: "decimal"
      currency:
        type: string
        enum: ["EUR", "USD"]
    required:
      - amount
      - currency
  ShowtimeResponse:
    type: object
    properties:
      id:
        type: string
        format: uuid
      movie_id:
        type: string
        format: uuid
      title:
        type: string
      date_start:
        type: string
        format: date-time
      date_end:
        type: string
        format: date-time
      price_override:
        $ref: "#/definitions/Price"
    required:
      - id
      - movie_id
      - title
      - date_start
      - date_end
  CreateShowtimeRequest:
    type: object
    properties:
      movie_catalog_id:
        type: string
        format: uuid
      date_start:
        type: string
        format: date-time
      date_end:
        type: string
        format: date-time
      price_override:
        $ref: "#/definitions/Price"
    required:
      - movie_catalog_id
      - date_start
      - date_end
...